requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
openai==1.3.7
redis==5.0.1
rq==1.15.1
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
pytz==2023.3
